Week 3 : 황지후 과제제출

+++ 황지후/3주차/3week_1.py
@@ -2,4 +2,4 @@ A = int(input("A값을 입력하시오 : "))
 B = int(input("B값을 입력하시오 : "))
 
 print ("{} {} {} {} {}".format(
-    (A + B), (A - B), (A * B), (A / B), (A % B)))
+    (A + B), (A - B), (A * B), int(A / B), (A % B)))
